requirements.txt
numpy<2.0.0
spacy>=3.7.0
networkx>=3.2
pandas>=2.0.0
langgraph>=0.0.40
langchain>=0.1.0
langchain-openai>=0.0.5
openai>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
pytest>=7.4.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
